apnoe/requirements.txt
streamlit~=1.35.0
pandas~=2.2.2
matplotlib~=3.9.0
PyYAML~=6.0.1
plotly
scikit-learn